requirements.txt
streamlit>=1.28.0
boto3>=1.28.0
Pillow>=9.0.0
opencv-python>=4.5.0
python-dotenv>=1.0.0
